requirements.txt
stylecloud==0.5.2
PySimpleGUI==4.60.4
stylecloud==0.5.2
Pillow==9.5.0
pandas==2.0.2
numpy==1.24.3
matplotlib==3.7.1
notebook==6.5.4